Add files via upload
In this code, I have done a new project of sentiment analysis in R. I applied the concept of sentiment analysis and implemented it over the dataset of Jane Austen’s books. I was able to delineate it through various visualizations after performing data wrangling on our data. I used a lexical analyzer – ‘bing’ in this instance of the project. Furthermore, I also represented the sentiment score through a plot and also made a visual report of wordcloud.